Move the features to plugins
- On On the fly polylines creation is now part of polylines_io_plugin
- On On the fly point sets creation is now part of xyz_plugin
