+++ Polyhedron/demo/Polyhedron/MainWindow.h
@@ -30,8 +30,6 @@ class QSortFilterProxyModel;
 
 namespace Ui {
   class MainWindow;
-  class Add_polylines_dialog;
-  class Add_point_set_dialog;
 }
 
 #include "Polyhedron_type_fwd.h"
@@ -268,19 +266,6 @@ protected Q_SLOTS:
   //Preferences edition
   //!Opens the Preferences dialog.
   void on_actionPreferences_triggered();
-  //!Opens a dialog to add polylines on the fly.
-  void on_actionAdd_polylines_triggered();
-  //!Opens a dialog to add a point set on the fly.
-  void on_actionAdd_point_set_triggered();
-  //!Adds a polyline
-  void addPolylineButton_clicked();
-  //!Adds a point set
-  void addPointSetButton_clicked();
-  //!Closes the dialog
-  void closePolylinesButton_clicked();
-  //!Closes the dialog
-  void closePointSetButton_clicked();
-
   // save as...
   //!Opens a dialog to save selected item if able.
   void on_actionSaveAs_triggered(); 
@@ -350,10 +335,6 @@ private:
   QSortFilterProxyModel* proxyModel;
   QTreeView* sceneView;
   Ui::MainWindow* ui;
-  Ui::Add_polylines_dialog *add_polydiagui;
-  Ui::Add_point_set_dialog *add_pointsetdiagui;
-  QDialog *add_polydiag;
-  QDialog *add_pointsetdiag;
   QVector<CGAL::Three::Polyhedron_demo_io_plugin_interface*> io_plugins;
   QMap<QString,QString> default_plugin_selection;
   // typedef to make Q_FOREACH work
