+++ Polyhedron/demo/Polyhedron/MainWindow.ui
@@ -53,9 +53,6 @@
     <addaction name="separator"/>
     <addaction name="actionLoad_Script"/>
     <addaction name="separator"/>
-    <addaction name="actionAdd_point_set"/>
-    <addaction name="actionAdd_polylines"/>
-    <addaction name="separator"/>
     <addaction name="actionQuit"/>
    </widget>
    <widget class="QMenu" name="menuEdit">
